requirements.txt
flask
flask-login
flask-sqlalchemy
scikit-learn
pandas
numpy
nltk
joblib
pytesseract
pillow
opencv-python
